+++ CHANGELOG.md
@@ -8,9 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 ### Changed
-- deps: update `quanta` to v0.12 and `erased-serde` to v0.4.
+- deps: update `quanta` to v0.12, `hyper` to v1 and `erased-serde` to v0.4.
 - context: slightly improve performance of time measurements.
 - utils: slightly improve performance of `RateLimiter`.
+- telemeter: handle only `GET /metrics` requests.
+- telemeter: rename the `address` config parameter to `listen` with alias to the old name.
+
+### Fixed
+- telemeter: avoid spawning tasks for a HTTP server, it improves common metrics.
 
 ## [0.2.0-alpha.11] - 2023-11-10
 ### Added